requirements.txt
openai>=1.14.0
python-dotenv
rich
flask